[IMP] improve filtering
Use the sublime text settings in addition to the default ripgrep filter
- binary_file_patterns
- file_exclude_patterns
- folder_exclude_patterns

Change the heuristic to use regex.

Here is a comparison of the new vs old ripgrep heuristic, on a big test project.

The old one is a bit random because it depends on the chars used for the search.

After 5 chars, the performances are very similar, but the new heuristic give better result.

Query   | Old    | New
--------|--------|-------
3 chars | ~220ms | ~720ms
4 chars | ~190ms | ~320ms
5 chars | ~180ms | ~290ms
6 chars | ~271ms | ~270ms
7 chars | ~210ms | ~240ms

We except people to not do a lot of "3 chars queries", so it has been changed.